# docs/swagger.yaml
basePath: /api/v1
...
host: localhost:3000
info:
  contact:
    email: support@datacollector.io
    name: API Support
    url: https://github.com/yourusername/datacollector
  description: Cryptocurrency market data collection service with OHLCV data, technical
    indicators, and job management.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: DataCollector API
  version: 1.0.5
paths:
...
  /exchanges/{id}/debug:
    get:
      consumes:
      - application/json
      description: Returns detailed debug information for a specific exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Debug information
          schema:
            additionalProperties: true
            type: object
      summary: Debug exchange
      tags:
      - Exchanges
  /exchanges/{id}/metadata:
    get:
      consumes:
      - application/json
      description: Returns metadata (timeframes, features, rate limits) for a specific
        exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Exchange metadata
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Exchange not found
          schema:
            additionalProperties: true
            type: object
      summary: Get exchange metadata
      tags:
      - Exchanges
  /exchanges/{id}/symbols:
    get:
      consumes:
      - application/json
      description: Returns all available trading symbols for a specific exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      - description: Search filter for symbols
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of symbols
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get exchange symbols
      tags:
      - Exchanges
  /exchanges/{id}/symbols/popular:
    get:
      consumes:
      - application/json
      description: Returns popular trading pairs that are available on the specific
        exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of popular symbols
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get popular symbols
      tags:
      - Exchanges
  /exchanges/{id}/symbols/validate:
    get:
      consumes:
      - application/json
      description: Validates if a trading symbol exists on a specific exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      - description: Symbol to validate (e.g., BTC/USDT)
        in: query
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Validation result with suggestions
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Missing symbol parameter
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Validate symbol
      tags:
      - Exchanges
    post:
      consumes:
      - application/json
      description: Validates multiple trading symbols at once on a specific exchange
      parameters:
      - description: Exchange ID
        in: path
        name: id
        required: true
        type: string
      - description: List of symbols to validate
        in: body
        name: request
        required: true
        schema:
          properties:
            symbols:
              items:
                type: string
              type: array
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Validation results
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Validate multiple symbols
      tags:
      - Exchanges
...
schemes:
- http
- https
swagger: "2.0"